Fix with-custom-reverse-proxy
