+++ examples/with-custom-reverse-proxy/package.json
@@ -8,9 +8,6 @@
     "react-dom": "16.2.0"
   },
   "devDependencies": {
-    "babel-preset-latest": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "babel-register": "^6.24.1",
     "cross-env": "^5.0.1",
     "http-proxy-middleware": "^0.17.4"
   },
